+++ cli/types.ts
@@ -110,7 +110,7 @@ export async function pushObj(
   } else if (typeEnding === "variable") {
     await pushVariable(workspace, p, befObj, newObj, plainSecrets);
   } else if (typeEnding === "flow") {
-    const flowName = p.split(".flow/")[0];
+    const flowName = p.split(".flow" + path.sep)[0];
     await pushFlow(workspace, flowName, flowName + ".flow", message);
   } else if (typeEnding === "resource") {
     await pushResource(workspace, p, befObj, newObj);
@@ -119,7 +119,9 @@ export async function pushObj(
   } else if (typeEnding === "schedule") {
     await pushSchedule(workspace, p, befObj, newObj);
   } else {
-    throw new Error("infer type unreachable");
+    throw new Error(
+      `The item ${p} has an unrecognized type ending ${typeEnding}`
+    );
   }
 }
 
@@ -150,7 +152,7 @@ export function getTypeStrFromPath(
   | "folder"
   | "app"
   | "schedule" {
-  if (p.includes(".flow/")) {
+  if (p.includes(".flow" + path.sep)) {
     return "flow";
   }
   const parsed = path.parse(p);