+++ frontend/src/lib/components/graph/FlowGraph.svelte
@@ -12,7 +12,8 @@
 		type Loop,
 		type Branch,
 		type NestedNodes,
-		type GraphModuleState
+		type GraphModuleState,
+		getStateColor
 	} from '.'
 	import { defaultIfEmptyString, encodeState } from '$lib/utils'
 	import { createEventDispatcher, onMount, setContext } from 'svelte'
@@ -159,6 +160,9 @@
 			const layered = layoutNodes(flatNodes)
 
 			nodes = layered.nodes
+
+			console.log(flatNodes, nodes)
+
 			let hfull = Math.max(layered.height, minHeight)
 			fullWidth = layered.width
 			height = fullSize ? hfull : Math.min(hfull, maxHeight ?? window.innerHeight * 1.5)
@@ -275,24 +279,6 @@
 			return [item.nodeEnd.id]
 		}
 		return []
-	}
-
-	function getStateColor(state: FlowStatusModule.type | undefined): string {
-		const isDark = document.documentElement.classList.contains('dark')
-		switch (state) {
-			case FlowStatusModule.type.SUCCESS:
-				return isDark ? '#059669' : 'rgb(193, 255, 216)'
-			case FlowStatusModule.type.FAILURE:
-				return isDark ? '#dc2626' : 'rgb(248 113 113)'
-			case FlowStatusModule.type.IN_PROGRESS:
-				return isDark ? '#f59e0b' : 'rgb(253, 240, 176)'
-			case FlowStatusModule.type.WAITING_FOR_EVENTS:
-				return isDark ? '#db2777' : 'rgb(229, 176, 253)'
-			case FlowStatusModule.type.WAITING_FOR_EXECUTOR:
-				return isDark ? '#ea580c' : 'rgb(255, 208, 193)'
-			default:
-				return isDark ? '#2e3440' : '#fff'
-		}
 	}
 
 	function getResultColor(): string {