+++ frontend/src/lib/stores.ts
@@ -43,7 +43,8 @@ export const dbClockDrift = writable<number | undefined>(undefined)
 export const isPremiumStore = writable<boolean>(false)
 export const starStore = writable(1)
 export const usersWorkspaceStore = writable<UserWorkspaceList | undefined>(undefined)
-export const superadmin = writable<String | false | undefined>(undefined)
+export const superadmin = writable<string | false | undefined>(undefined)
+export const lspTokenStore = writable<string | undefined>(undefined)
 export const userWorkspaces: Readable<
 	Array<{
 		id: string